🚚 refactor: APIの型定義を共通ファイルに切り出し
