+++ src/app/_features/BlogPostList/BlogPostListContainer.tsx
@@ -1,37 +1,12 @@
 import { BlogPostListUI } from "./BlogPostListUI";
-import { client } from "@libs/microcms";
-
-// TODO: typeの場所移動すること
-// TODO: 呼び出し関数用のディレクトリを作成すること
-type Tag = {
-	id: string;
-	name: string;
-};
-
-type Thumbnail = {
-	url: string;
-};
-
-export type BlogPost = {
-	id: string;
-	title: string;
-	tags: Tag[];
-	createdAt: string;
-	thumbnail?: Thumbnail;
-};
-
-type Response = {
-	contents: BlogPost[];
-	totalCount: number;
-	offset: number;
-	limit: number;
-};
+import { client } from "@/libs/microcms";
+import type { BlogResponse } from "@/libs/microcms/blogs/types";
 
 type BlogPostListContainerProps = {
 	page: number;
 };
 
-async function getBlogPosts(page: number): Promise<Response> {
+async function getBlogPosts(page: number): Promise<BlogResponse> {
 	const data = await client.get({
 		endpoint: "blogs",
 		queries: {
